Revert "Support `import_map.json[c]` in denops plugin directory"
