+++ deno.jsonc
@@ -13,8 +13,7 @@
   },
   "exclude": [
     ".coverage/",
-    "tests/denops/testdata/no_check/",
-    "tests/denops/testdata/with_import_map/"
+    "tests/denops/testdata/no_check/"
   ],
   // NOTE: Import maps should only be used from test modules.
   "imports": {
