--- tests/denops/testdata/with_import_map/helper.ts
@@ -1,5 +1,0 @@
-export const greeting = "Hello from mapped import!";
-
-export function getMessage(name: string): string {
-  return `Import map works for ${name}!`;
-}
