--- tests/denops/testdata/with_import_map/import_map.jsonc
@@ -1,10 +1,0 @@
-{
-  // This is a JSONC import map with comments
-  "imports": {
-    "@test/": "../",
-    // Map to a specific module
-    "@test/lib": "../dummy_valid_plugin.ts",
-    // Map to the helper module
-    "@test/helper": "./helper.ts"
-  }
-}
