--- tests/denops/testdata/with_import_map/plugin_with_import_map.ts
@@ -1,13 +1,0 @@
-import type { Entrypoint } from "jsr:@denops/core@^7.0.0";
-import { getMessage, greeting } from "@test/helper";
-
-export const main: Entrypoint = async (denops) => {
-  denops.dispatcher = {
-    test: async () => {
-      const message = getMessage("test-plugin");
-      await denops.cmd(`echo '${message}'`);
-      return greeting;
-    },
-  };
-  await denops.cmd("echo 'Import map plugin initialized'");
-};
